+++ packages/git-mob-core/src/git-mob-api/git-config.ts
@@ -1,11 +1,12 @@
 import { getConfig, setConfig } from './exec-command.js';
+import { resolveGitMessagePath } from './resolve-git-message-path.js';
 
 export async function getLocalCommitTemplate() {
   return getConfig('--local commit.template');
 }
 
 export async function getGlobalCommitTemplate() {
-  return getConfig('--global commit.template');
+  return (await getConfig('--global commit.template')) || resolveGitMessagePath();
 }
 
 export async function getGitUserName() {